Create README.md
# Rabbit-Hole
This is an artistic-research psychedelic project aimed at revealing the deepest psychedelic experiences to humanity with the help of artificial intelligence animations. So if you're interested in seeing the full depth of the rabbit hole in ultra-high definition, get ready.

"Morpheus: This is your last chance. After this, there is no turning back. You take the blue pill - the story ends, you wake up in your bed and believe whatever you want to believe. You take the red pill - you stay in Wonderland and I show you how deep the rabbit hole goes." Matrix, 1999

Mission:
The Rabbit Hole project was created to secure the strongest psychedelic experiences created by artist in collaboration with artificial intelligence in the blockchain. People who cannot get such an experience can get it through these animations and thus expand their consciousness. So this project contributes to the evolution of humanity to a higher level of mind.

Goals:
- Create a large database of psychedelic animations in high quality on Optimism.
- Bringing more people into the world of psychedelic art.

Roadmap:
1. Writing scripts/prompts for animations based on real-life psychedelic experiences.
2. Selection of one of a psychedelic musicians and collaborations.
3. Creating camera motion code in sync with music for Deforum extension in Stable Diffusion.
4. Generating animarion.
5. Mint on Optimism in ultra-high resolution.
6. Promote this animations in social media, exhibitions and conferences.
7. Create more animations (3-5 animations  3 minutes each per month).
8. Complete animations in long time movies and DJ sets for long time expiriences.

Budget:
Powerful PC with GPU RTX 4090ti + Monitor: 4000-5000$
Marketing and events (exhibitions, attending conferences): 2000$
Artist compensation: 0
Total: 7000$

Conclusion:
This project can change a lot of consciousness in people open to the mind of evolution and will make history, or one of the histories in the infinite space of the quantum multiverse.